Move Generic out of legacy folder
